output matches "git-sizer -j"
Different stages implemented to know non-reachable objects and avoid considering them in the statistics.

Threads leveraged in two different places: when reading and storing objects information from the database, and when setting reachable objects.

Windows build fix: std::max() was conflicting with macro definition in windows.h. Fixed using template syntax std::max<type>() so that it's not confused with the macro.